% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichr.R
\name{enrichr}
\alias{enrichr}
\title{Connect to the enrichr API to calculate enrichments, then plot them in a cellplot}
\usage{
enrichr(result, gene_set_library = "ChEA_2015", title = "", onlysig = F,
  maxNumberOfTerms = 15, fileroot = "", ...)
}
\arguments{
\item{result}{Dataframe containing differential expression result. Should have the column "external_gene_id", which contains the official gene name, and the column "log2FoldChange".}

\item{gene_set_library}{Character. Which Enrichr gene set library should be used? Names are listed on the Enrichr website: http://amp.pharm.mssm.edu/Enrichr/#stats}

\item{title}{Title for the plot}

\item{onlysig}{Boolean. Should it plot only the significant results?}

\item{maxNumberOfTerms}{Numeric. How many terms should be shown? Is overruled by setting onlysig to TRUE.}

\item{fileroot}{This function saves its results to a file, to prevent repeatedly sending the same query to Enrichr. This speeds up the process if you are for example knitting a markdown file later.}
}
\value{
The enrichment results calculated by Enrichr.
}
\description{
Connect to the enrichr API to calculate enrichments, then plot them in a cellplot
}
\examples{
enrichr(SignificantGenes)
}
